+++ src/index/ix_index_handle.h
@@ -51,8 +51,8 @@ inline int ix_compare(const char* a, const char* b, ColType type, int col_len) {
         return (ia < ib) ? -1 : ((ia > ib) ? 1 : 0);
     }
     case TYPE_FLOAT: {
-        double fa;
-        double fb;
+        float fa;
+        float fb;
         memcpy(&fa, a, sizeof(fa));
         memcpy(&fb, b, sizeof(fb));
         return (fa < fb) ? -1 : ((fa > fb) ? 1 : 0);
